# services/qualitative-service/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GenNet Qualitative Service API",
    "version": "2.0.0",
    "description": "API for qualitative modeling and SMBioNet integration"
  },
  "servers": [
    {
      "url": "https://api.gennet.example.com/api/v1/qualitative",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/ctl/verify": {
      "post": {
        "summary": "Verify CTL formula",
        "description": "Verify the syntax and semantics of a CTL formula",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CTLFormula"
              },
              "example": {
                "formula": "AG(p -> AF q)",
                "description": "Property description"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CTLVerificationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/parameters/generate": {
      "post": {
        "summary": "Generate parameters",
        "description": "Generate K-parameters from CTL formula",
        "parameters": [
          {
            "name": "network_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParameterGenerationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterGenerationResult"
                }
              }
            }
          }
        }
      }
    },
    "/health/live": {
      "get": {
        "summary": "Liveness probe",
        "description": "Kubernetes liveness probe endpoint",
        "responses": {
          "200": {
            "description": "Service is alive"
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "summary": "Readiness probe",
        "description": "Kubernetes readiness probe endpoint",
        "responses": {
          "200": {
            "description": "Service is ready"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CTLFormula": {
        "type": "object",
        "required": ["formula"],
        "properties": {
          "formula": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000,
            "description": "CTL formula string"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "description": "Optional description"
          }
        }
      },
      "CTLVerificationResult": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "formula": {
            "type": "string"
          },
          "parsed_structure": {
            "type": "object"
          }
        }
      },
      "ParameterGenerationRequest": {
        "type": "object",
        "required": ["formula"],
        "properties": {
          "formula": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "network_structure": {
            "type": "object"
          }
        }
      },
      "ParameterGenerationResult": {
        "type": "object",
        "properties": {
          "parameters": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "count": {
            "type": "integer"
          },
          "network_id": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}